refactor some functions, moved to utils.py
