+++ wsgi.py
@@ -4,10 +4,8 @@ from flask import Flask, Response, request
 from flask_mysqldb import MySQL
 from umichemail import send_mail
 from dotenv import load_dotenv
+from utils import *
 
-import hmac
-import hashlib
-import socket
 import logging
 import os
 
@@ -25,43 +23,6 @@ secret = os.environ.get('SECRET', 'none')
 logging.basicConfig(level=log_level)
 
 
-def reverseLookup(IP):
-    try:
-        return socket.gethostbyaddr(IP)[0].lower()
-    except Exception:
-        return IP
-
-
-def test_mysql_connection():
-    try:
-        conn = mysql.connection.cursor()
-        if conn:
-            #app.logger.debug("Mysql test connection successful")
-            return True
-        else:
-            app.logger.error("Mysql test connection failed")
-            return False
-    except Exception as e:
-        app.logger.error("Mysql connection test exception {}".format(e))
-        return False
-
-
-def test_mysql_query():
-    try:
-        cursor = mysql.connection.cursor()
-        cursor.execute("SELECT DATABASE()")
-        result = cursor.fetchall()
-        if result:
-            #app.logger.debug("Mysql %s test query successful", result)
-            return True
-        else:
-            app.logger.error("Mysql test query failed")
-            return False
-    except Exception as e:
-        app.logger.error("Mysql query test exception %s", e.__cause__)
-        return False
-
-
 @app.route('/hello')
 def hello_world():
     if request.remote_addr.startswith('67.149'):
@@ -74,7 +35,7 @@ def hello_world():
 
 @app.route('/health')
 def test_mysql():
-    if test_mysql_connection() and test_mysql_query():
+    if test_mysql_connection(app, mysql) and test_mysql_query(app, mysql):
         return "MySQL connection and query test successful."
     else:
         return "MySQL connection or query test failed.", 500
@@ -85,73 +46,59 @@ def zoomphone_registration():
     if not request.is_json:
         return Response("Invalid request", status=400)
 
-    zm_request_timestamp = request.headers.get('x-zm-request-timestamp')
-    message = f'v0:{zm_request_timestamp}:{request.data.decode("utf-8")}'
-    app.logger.debug("string to use for hmac: %s", message)
-    our_sig = 'v0=' + hmac.new(secret.encode("utf-8"),
-                               message.encode("utf-8"),
-                               hashlib.sha256 ).hexdigest()
-    app.logger.debug("our_sig: %s", our_sig)
+    if not validate_request(app, request, secret):
+        print("could not validate request hash from host: ",
+              reverseLookup(request.remote_addr))
+        return Response("Access denied", 401)
+    else:
+        app.logger.debug("device registration webhook received from: %s",
+                        reverseLookup(request.remote_addr))
+        app.logger.debug("%s", request.data.decode())
+        if request.is_json:
+            data = request.get_json()
+            if data:
+                device_id = data['payload']['object']['device_id']
+                mac_address = data['payload']['object']['mac_address']
+                app.logger.info("got webhook for device_id: {} mac_address: {}".format(
+                    device_id, mac_address))
 
-    zm_signature = request.headers.get('x-zm-signature', type=str)
-    app.logger.debug("x-zm-signature: %s", zm_signature)
-    if zm_signature == our_sig:
-        app.logger.debug("message hmac signatures match")
-
-    token = request.headers.get('authorization', type=str)
-    #if token == app.config["ZOOM_TOKEN"]:
-    app.logger.debug("device registration webhook received from: %s",
-                    reverseLookup(request.remote_addr))
-    app.logger.debug("%s", request.data.decode())
-    if request.is_json:
-        data = request.get_json()
-        if data:
-            device_id = data['payload']['object']['device_id']
-            mac_address = data['payload']['object']['mac_address']
-            app.logger.info("got webhook for device_id: {} mac_address: {}".format(
-                device_id, mac_address))
-
-            ts = datetime.now(tz=tz.gettz('America/Detroit'))
-            sql_vals = (device_id, ts.strftime(
-                '%Y-%m-%d %H:%M:%S'), "ph_" + mac_address + "%")
-            sql = "UPDATE ZoomPhoneNameFloorRoom SET deviceId = %s, stamp = %s WHERE PhoneName LIKE %s"
-            app.logger.debug(
-                'Attempting mysql update with: {}'.format(sql_vals))
-            try:
-                cur = mysql.connection.cursor()
-                if not cur:
-                    msg = 'Unable to connect to MySQL DB: {}:{}'.format(
-                        app.config["MYSQL_HOST"], app.config["MYSQL_DB"])
-                    app.logger.error(msg)
-                    send_mail(message=msg, subject='ERROR: Zoom webhook Mysql connection',
-                                from_address=mail_from, to_address=mail_to)
-                    return Response("Database connection failed", headers='Retry-After: 300', status=503)
-                else:
-                    result = cur.execute(sql, sql_vals)
-                    mysql.connection.commit()
-                    if result == 0:
-                        sql = "INSERT INTO ZoomPhoneNameFloorRoom ( PhoneName, deviceId, stamp ) VALUES(%s, %s, %s);"
-                        sql_vals = ("ph_" + mac_address, device_id, ts.strftime('%Y-%m-%d %H:%M:%S'))
+                ts = datetime.now(tz=tz.gettz('America/Detroit'))
+                sql_vals = (device_id, ts.strftime(
+                    '%Y-%m-%d %H:%M:%S'), "ph_" + mac_address + "%")
+                sql = "UPDATE ZoomPhoneNameFloorRoom SET deviceId = %s, stamp = %s WHERE PhoneName LIKE %s"
+                app.logger.debug(
+                    'Attempting mysql update with: {}'.format(sql_vals))
+                try:
+                    cur = mysql.connection.cursor()
+                    if not cur:
+                        msg = 'Unable to connect to MySQL DB: {}:{}'.format(
+                            app.config["MYSQL_HOST"], app.config["MYSQL_DB"])
+                        app.logger.error(msg)
+                        send_mail(message=msg, subject='ERROR: Zoom webhook Mysql connection',
+                                    from_address=mail_from, to_address=mail_to)
+                        return Response("Database connection failed", headers='Retry-After: 300', status=503)
+                    else:
                         result = cur.execute(sql, sql_vals)
                         mysql.connection.commit()
-                        app.logger.debug("SQL insert result: %s", result)
-                        msg = 'device registration webhook was received but no existing entries in DB matched for phone with MAC: {}'.format(
-                            mac_address)
-                        app.logger.error(msg)
-                        send_mail(message=msg, subject='Notice: Zoom device registration, missing DB entry',
-                                    from_address=mail_from, to_address=mail_to)
-                cur.close()
-            except Exception as e:
-                mysql.connection.rollback()
-                cur.close()
-                send_mail(message=str(e), subject='SQL Exception occurred',
-                            from_address=mail_from, to_address=mail_to)
-                app.logger.error("SQL Exception occurred: ", str(e.with_traceback))
-    return Response("", 200)
-    #else:
-    #    print("invalid auth token: ", token, "from host: ",
-    #          reverseLookup(request.remote_addr))
-    #    return Response("Access denied", 401)
+                        if result == 0:
+                            sql = "INSERT INTO ZoomPhoneNameFloorRoom ( PhoneName, deviceId, stamp ) VALUES(%s, %s, %s);"
+                            sql_vals = ("ph_" + mac_address, device_id, ts.strftime('%Y-%m-%d %H:%M:%S'))
+                            result = cur.execute(sql, sql_vals)
+                            mysql.connection.commit()
+                            app.logger.debug("SQL insert result: %s", result)
+                            msg = 'device registration webhook was received but no existing entries in DB matched for phone with MAC: {}'.format(
+                                mac_address)
+                            app.logger.error(msg)
+                            send_mail(message=msg, subject='Notice: Zoom device registration, missing DB entry',
+                                        from_address=mail_from, to_address=mail_to)
+                    cur.close()
+                except Exception as e:
+                    mysql.connection.rollback()
+                    cur.close()
+                    send_mail(message=str(e), subject='SQL Exception occurred',
+                                from_address=mail_from, to_address=mail_to)
+                    app.logger.error("SQL Exception occurred: ", str(e.with_traceback))
+        return Response("", 200)
 
 
 if __name__ == '__main__':
